requirements.txt
# Core web framework
flask==3.0.3
flask-cors==4.0.1

# Scheduler
apscheduler==3.10.4

# AI Model SDKs
anthropic==0.28.0
openai==1.35.7
google-generativeai==0.7.2

# Stock data
yfinance==0.2.40

# Database
# (SQLite is built into Python — no extra package needed)

# Email / utilities
python-dotenv==1.0.1

# HTTP / JSON
requests==2.32.3

# Data processing
pandas==2.2.2

# Production WSGI server (for Railway/Render)
gunicorn==22.0.0
